requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
python-dotenv
faiss-cpu
